+++ apps/payments/create-react-app/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/client-sdk-react-ui-starter
 
+## 1.2.15
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.12
+
+## 1.2.14
+
+### Patch Changes
+
+-   Updated dependencies [36204ce]
+    -   @crossmint/client-sdk-react-ui@1.3.11
+
 ## 1.2.13
 
 ### Patch Changes