+++ apps/wallets/smart-wallet/next/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/client-sdk-smart-wallet-next-starter
 
+## 0.1.8
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.12
+
+## 0.1.7
+
+### Patch Changes
+
+-   Updated dependencies [36204ce]
+    -   @crossmint/client-sdk-react-ui@1.3.11
+
 ## 0.1.6
 
 ### Patch Changes