+++ packages/client/ui/react-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @crossmint/client-sdk-react-ui
 
+## 1.3.12
+
+### Patch Changes
+
+-   Updated dependencies [f8f0c66]
+    -   @crossmint/client-sdk-base@1.2.4
+    -   @crossmint/client-sdk-smart-wallet@0.1.11
+
+## 1.3.11
+
+### Patch Changes
+
+-   36204ce: fix error handling issue
+-   Updated dependencies [4c5d88b]
+    -   @crossmint/client-sdk-smart-wallet@0.1.10
+
 ## 1.3.10
 
 ### Patch Changes